fix: guard empty-batch loss (n<=0) and widen MT-threshold to size_t
Two correctness fixes surfaced by the QA review pass:

- tk_loss / tk_train_step_f32 returned NaN for n<=0: inv=1/0=+inf, the
  sum loop never runs (sum=0), and 0*inf=NaN silently poisons the loss.
  An empty batch/layer now returns a clean 0.0f.

- The multithread-dispatch work estimate used (long)out_dim*in_dim.
  long is 32-bit on LLP64 (Windows, a release-build target), so the
  comparison could overflow for very large layers and mis-decide
  threading. Widened to size_t, matching the (size_t) casts already used
  for the row-index arithmetic. LP64 (macOS/Linux) behaviour unchanged.

Verified: make test (bf16) + make testbp still pass.

+++ src/ops.c
@@ -178,7 +178,7 @@ void tk_linear_forward(const tk_scalar_t *restrict W,
                        float *restrict y,
                        int out_dim, int in_dim,
                        tk_activation_t act) {
-    if ((long)out_dim * in_dim >= TK_MT_MIN_WORK && tk_num_threads() > 1) {
+    if ((size_t)out_dim * in_dim >= TK_MT_MIN_WORK && tk_num_threads() > 1) {
         tk_gemv_ctx c = { W, x, bias, y, in_dim, act };
         tk_parallel_for(out_dim, gemv_worker, &c);   /* split rows across the pool */
     } else {
